+++ src/impl/v2/index.js
@@ -3,8 +3,9 @@
  */
 
 const { JSONPath: jsonPath } = require('jsonpath-plus'),
-    { setNoAdditionalProperties } = require('../service/disallow-additional-properties'),
-    cloneDeep = require('lodash.clonedeep');
+    cloneDeep = require('lodash.clonedeep'),
+    { setAllPropertiesRequired } = require('../service/all-properties-required'),
+    { setNoAdditionalProperties } = require('../service/no-additional-properties');
 
 // CONSTANTS
 
@@ -50,11 +51,13 @@ function buildValidationMap(pathsExamples) {
  * The passed spec won't be modified. If a modification happens, a modified copy will be returned.
  * @param {Object}  openapiSpec                     The OpenAPI-spec as JSON-schema
  * @param {boolean} [noAdditionalProperties=false]  Don't allow properties that are not defined in the schema
+ * @param {boolean} [allPropertiesRequired=false]   Make all properties required
  * @return {Object} The prepared OpenAPI-spec
  */
-function prepare(openapiSpec, { noAdditionalProperties } = {}) {
+function prepare(openapiSpec, { noAdditionalProperties, allPropertiesRequired } = {}) {
     const openapiSpecCopy = cloneDeep(openapiSpec);
     noAdditionalProperties && setNoAdditionalProperties(openapiSpecCopy, getJsonPathsToExamples());
+    allPropertiesRequired && setAllPropertiesRequired(openapiSpecCopy, getJsonPathsToExamples());
     return openapiSpecCopy;
 }