+++ src/validator.js
@@ -6,6 +6,7 @@ const { JSONPath: jsonPath } = require('jsonpath-plus'),
     JsonPointer = require('json-pointer'),
     Ajv = require('ajv-draft-04'),
     FormatValidator = require('ajv-oai/lib/format-validator');
+const { validateDateTime } = require('./validators/date-time');
 
 const PROP__ID = '$id',
     JSON_PATH__REFS = '$..\$ref',
@@ -118,4 +119,5 @@ function _addFormatValidators(validator) {
     validator.addFormat('float', { type: 'number', validate: FormatValidator.float });
     validator.addFormat('double', { type: 'number', validate: FormatValidator.double });
     validator.addFormat('byte', { type: 'string', validate: FormatValidator.byte });
+    validator.addFormat('date-time', { type: 'string', validate: validateDateTime });
 }