+++ test/data/output-help.js
@@ -13,6 +13,7 @@ Options:
   -m, --mapping-filepath <mapping-filepath>  file path to map, containing schema-paths as key and the file-path(s) to examples as value. If wildcards are used, the parameter has to be put in quotes.
   -c, --cwd-to-mapping-file                  changes to the directory of the mapping-file, before resolving the example's paths. Use this option, if your mapping-files use relative paths for the examples
   -n, --no-additional-properties             don't allow properties that are not described in the schema
+  -r, --all-properties-required              make all the properties in the schema required
   -o, --ignore-formats <ignored-formats...>  Datatype formats to ignore (to prevent "unknown format" errors.)
   -h, --help                                 display help for command